refactor(web): extract custom-endpoints routes into web_routers/custom_endpoints (web_server.py god-file slice R3-C1)
